Fix WITH_APU_CRYPTO check on 32-bit Linux platform
All of the apr flags needed to compile APU_HAVE_CRYPTO check
in configure.ac aren't passed.  While this works fine for 64-bit
machines (because _LARGEFILE64_SOURCE is already defined),
this does not work on 32-bit.  This in-turn breaks the apr_off_t
definition in apr.h.

By passing along the apr --cflags and --cppflags to compile,
32-bit machines will allow WITH_APU_CRYPTO to be set if there's
support for it.